producto service terminado

+++ msvc-producto/src/main/java/com/ecomarket/msvc/producto/models/Producto.java
@@ -30,6 +30,6 @@ public class Producto {
     private String descripcion;
 
     @Column(nullable = false)
-    private Long stock;
+    private int stock;
 
 }
